requirements.txt
absl-py==0.8.1
aim==2.1.4
aimrecords==0.0.7
anytree==2.8.0
astor==0.8.0
axial-positional-embedding==0.2.1
base58==2.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==7.1.2
contrastive-learner==0.1.0
cycler==0.10.0
decorator==4.4.1
docker==4.4.1
dominate==2.4.0
easydict==1.9
fire==0.3.1
flopth==0.0.1
future==0.18.2
gast==0.3.2
gitdb==4.0.5
GitPython==3.1.11
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.9.0
idna==2.8
imageio==2.6.1
inplace-abn==1.0.7
joblib==0.15.1
jsonpatch==1.24
jsonpointer==2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kornia==0.4.1
linear-attention-transformer==0.15.2
linformer==0.2.1
lmdb==0.98
local-attention==1.1.1
Markdown==3.1.1
matplotlib==3.1.2
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
natsort==7.1.1
networkx==2.4
neuralgym==0.0.1
nibabel==3.0.0
numpy==1.17.2
olefile==0.46
opencv-python==4.1.2.30
packaging==20.4
pandas==1.0.5
Pillow==6.2.0
product-key-memory==0.1.10
protobuf==3.14.0
psutil==5.8.0
ptflops==0.6.2
py==1.10.0
pyarrow==0.17.1
pycparser==2.19
pyparsing==2.4.5
pyrser==0.2.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.22.0
retry==0.9.2
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.1.0
six==1.13.0
sklearn==0.0
smmap==3.0.4
stylegan2-pytorch==1.5.6
tabulate==0.8.7
tensorboard==1.14.0
tensorboardX==1.9
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tflearn==0.3.2
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
torch==1.2.0
torchfile==0.1.0
torchstat==0.0.7
torchtoolbox==0.1.4.1
torchvision==0.2.1
tornado==6.0.3
tqdm==4.46.1
urllib3==1.25.7
vector-quantize-pytorch==0.1.0
visdom==0.1.8.9
websocket-client==0.56.0
Werkzeug==0.16.0
wget==3.2
wrapt==1.11.2
yacs==0.1.8
